import the provider registries from server/lib instead of copying them into client/src/utils
The client/src/lib mirrors were converted to re-exports of their pure
server/lib leaf in #6364; the provider-domain mirrors in client/src/utils
were outside that sweep and still carried hand copies of the gateway
registry, the harness id→label table, the effort ladders and sentinels,
the local-model capability regexes, the context-window ladder and the
caller-mode policy — each guarded by a parity test that read the client
file as text and asked a human to port the change. That is recurring
churn, not a guard: the gpt-6 `minimal` gate (f4df11740) had to land the
same rule in both files AND a new row in the mirror test.

Every one of those server modules is already browser-safe (the client
imports 40 other server/lib leaves), so:

- providerGateways / localModelHeuristics re-export the server leaf
  (`isGatewayBackedProvider` moves to the server leaf too);
  providerHarnesses names a harness through `harnessById`;
  providerSelection reads `CALLER_MODE_POLICIES` directly.
- providerModels re-exports the sentinels, ladders and the Antigravity
  split. Its `effortLevelsForProvider` now delegates to the server's and
  adds only the rung the browser needs — the ladder the server publishes
  on a sanitized provider inventory — keeping the server's null final for
  Antigravity; `resolveCliEffort` clamps through a new exported server
  `clampEffortToLadder` instead of restating the rank walk.
- The context-window ladder moves out of services/stageRunner.js into a
  new pure leaf lib/providerContextWindows.js with its own test;
  stageRunner imports the rungs and keeps `effectiveContextWindow`. The
  leaf keys the vendor windows on the shared `isXProvider` predicates
  rather than a second hand-rolled id/basename test, so a command spelled
  `antigravity` now gets the Gemini window like the effort ladder already
  gave it.
- Delete the three parity suites whose only assertion was copy equality
  (providerHarnesses.parity, providerModels.mirror, stageRunner.mirror),
  shrink providerGateways.parity and localModelHeuristics.mirror to the
  aiToolkit copies the vendoring rule still requires, drop the
  now-tautological table comparison from the client mode-policy test, and
  remove mirrorParity's regex-alternation comparison, which those suites
  were the only consumers of.
- Fix client/src/lib/README.md's pointers at the two import guards, which
  named files that do not exist.
