+++ client/src/utils/localModelHeuristics.js
@@ -4,10 +4,11 @@
  * the union with the backend's own capability report where the server could
  * probe one (`useToolUseModelIds` / `useVisionModelIds`).
  *
- * Browser MIRROR of `server/lib/localModelHeuristics.js` — the regexes are
- * inlined here because the browser cannot import server code, and
- * `server/lib/localModelHeuristics.mirror.test.js` reads this file as TEXT and
- * fails when a pattern stops matching the same ids as the server's.
+ * The family predicates (`isEmbeddingModel` / `isVisionModel` /
+ * `isToolUseModel`) are re-exported from the pure leaf
+ * `server/lib/localModelHeuristics.js`. They take a bare id string (the case
+ * every caller here has) or a model card, whose explicit capability metadata
+ * outranks the id regex.
  *
  * Re-exported by `./providers.js` for existing `utils/providers` imports.
  */
@@ -15,69 +16,23 @@
 import { isLocalInstanceProvider, localBackendForProvider } from './providerEndpoints.js';
 import { filterSelectableModels } from './providerModels.js';
 import { commandBasename, isOllamaBackedProvider } from './providerTypes.js';
+import { isEmbeddingModel, isToolUseModel, isVisionModel } from '../../../server/lib/localModelHeuristics.js';
 
-/**
- * Embedding-only model detector — mirror of `isEmbeddingModel` in
- * server/lib/localModelHeuristics.js. Keep the two regexes in lockstep (the
- * server lib can't be imported here). Used to keep embedding models (e.g.
- * `nomic-embed-text`) out of generation/chat model pickers.
- * @param {string} id
- * @returns {boolean}
- */
-export const isEmbeddingModel = (id) =>
-  typeof id === 'string' && id.length > 0 &&
-  // Mirror of EMBEDDING_RE in server/lib/localModelHeuristics.js — keep in lockstep.
-  // `embeddinggemma` needs its own alternative: the anchored `embedding` marker
-  // requires a separator after it, and that id glues the family straight on.
-  // `minilm` / `paraphrase-multilingual` carry no `embed` marker at all —
-  // `all-minilm` and `paraphrase-multilingual` are Ollama embedding models that
-  // would otherwise be offered in a chat/generation picker.
-  /(?:^|[-_/:])(?:embed|embedding|bge|nomic|mxbai|gte|e5|snowflake-arctic-embed)(?:[-_/:]|$)|text-embedding|embeddinggemma|minilm|paraphrase-multilingual/i.test(id);
-
-/**
- * Vision-capable (multimodal) model detector — mirror of `isVisionModel` in
- * server/lib/localModelHeuristics.js (id-regex branch only). Keep the regex in
- * lockstep with the server. Used to flag/select vision models in the LoRA
- * caption picker. The server prefers explicit backend capability metadata
- * (`vision: true` on the model card); use that field when you have it and fall
- * back to this for bare id strings.
- * @param {string} id
- * @returns {boolean}
- */
-export const isVisionModel = (id) =>
-  typeof id === 'string' && id.length > 0 &&
-  // Mirror of VISION_RE in server/lib/localModelHeuristics.js — keep in lockstep.
-  /(?:^|[-_/:])vision(?:[-_/:.]|$)|(?:^|[-_/:])vl(?:\d|[-_/:.]|$)|qwen[\d.]*-?vl|(?:^|[-_/:])gemma-?[34]|llava|bakllava|moondream|minicpm-?v|pixtral|smolvlm|internvl|cogvlm|glm-?4v|phi-?3\.5?-vision|phi-?4-multimodal|got-ocr|idefics|fuyu|paligemma|kosmos|nanollava/i.test(id);
+export { isEmbeddingModel, isToolUseModel, isVisionModel } from '../../../server/lib/localModelHeuristics.js';
 
 /**
  * Whether a CLI-type provider can read an image file (its CLI accepts a
- * vision attachment). Mirror of `isVisionCapableCliProvider` in
- * server/lib/localModelHeuristics.js — keyed on command basename so a
- * renamed/path-configured Claude or Codex still qualifies. API providers
- * return false here; use `visionLocalModelFilter` for their model lists.
+ * vision attachment). The client's own take on `isVisionCapableCliProvider`
+ * in server/lib/localModelHeuristics.js: keyed on the command basename so a
+ * path-configured Claude or Codex still qualifies here, where the server
+ * matches the bare command exactly. API providers return false; use
+ * `visionLocalModelFilter` for their model lists.
  * @param {{type?:string, command?:string}|null|undefined} provider
  * @returns {boolean}
  */
 export const isVisionCapableCliProvider = (provider) =>
   provider?.type === 'cli'
   && (commandBasename(provider.command) === 'codex' || commandBasename(provider.command) === 'claude');
-
-/**
- * Tool-use (function-calling) capable model detector — mirror of `isToolUseModel`
- * in server/lib/localModelHeuristics.js (and the TOOL_USE_RE inlined in
- * server/lib/aiToolkit/providers.js). Keep all three in lockstep (the server libs
- * can't be imported here) — server/lib/localModelHeuristics.mirror.test.js reads
- * this file as text and fails when the patterns stop matching the same ids. Ollama's /api/show `tools` capability is authoritative
- * when known; this id regex is the fallback for bare model-id strings. The CoS
- * agent harness depends on reliable tool-calling, so only these families should
- * be selectable for a local-model-backed coding provider.
- * @param {string} id
- * @returns {boolean}
- */
-export const isToolUseModel = (id) =>
-  typeof id === 'string' && id.length > 0 &&
-  // Mirror of TOOL_USE_RE in server/lib/localModelHeuristics.js — keep in lockstep.
-  /qwen|llama-?3\.[1-9]|llama-?4|mistral|mixtral|ministral|codestral|devstral|magistral|command-?r|command-?a|north-mini-code|firefunction|functionary|watt-tool|hermes|functiongemma|glm-?4|granite-?[34]|(?:^|[-_/:])gemma-?4|gpt-oss|nemotron|olmo-?3|lfm2|ornith|muse-glimmer|nex-n2|smollm2|dflash|deepseek-v3|deepseek-r1|deepseek-v4/i.test(id);
 
 /**
  * Agent-picker tool-use annotation for a model id. Agent / CoS tasks (the CD
