+++ server/lib/index.js
@@ -424,6 +424,7 @@ export * from './fleetQuotas.js';
 export * from './harnessOutput.js';
 export * from './providerGateways.js';
 export * from './providerHarnesses.js';
+export * from './providerContextWindows.js';
 export * from './providerConnections.js';
 export * from './providerGraphPreview.js';
 export * from './providerGraphRecords.js';
