+++ server/lib/mirrorParity.test.js
@@ -3,8 +3,6 @@ import {
   stripCommentsAndNormalize,
   extractDeclaration,
   compareDeclaration,
-  compareRegexDeclaration,
-  regexAlternationSource,
 } from './mirrorParity.js';
 
 describe('stripCommentsAndNormalize', () => {
@@ -97,71 +95,5 @@ describe('compareDeclaration', () => {
     const { clientDecl, match } = compareDeclaration(server, 'const other = 1;', 'f');
     expect(clientDecl).toBeNull();
     expect(match).toBe(false);
-  });
-});
-
-describe('compareRegexDeclaration', () => {
-  // The shape this exists for: the server spells a capability regex as an array
-  // of per-alternative strings (so each alternative can carry its own comment)
-  // while the browser copy inlines the same pattern as one literal.
-  const server = "const TOOL_USE_RE = new RegExp([\n  'qwen',\n  'llama-?3\\\\.[1-9]',\n].join('|'), 'i');\n";
-  const client = "export const isToolUseModel = (id) =>\n  /qwen|llama-?3\\.[1-9]/i.test(id);\n";
-
-  it('matches the array form against the inline literal it compiles to', () => {
-    const { serverSource, match } = compareRegexDeclaration(server, client, 'TOOL_USE_RE', 'isToolUseModel');
-    expect(serverSource).toBe('qwen|llama-?3\\.[1-9]');
-    expect(match).toBe(true);
-  });
-
-  it('reports a mismatch when one side gains an alternative', () => {
-    const drifted = "export const isToolUseModel = (id) =>\n  /qwen|llama-?3\\.[1-9]|newfamily/i.test(id);\n";
-    expect(compareRegexDeclaration(server, drifted, 'TOOL_USE_RE', 'isToolUseModel').match).toBe(false);
-  });
-
-  it('treats an escaped and a bare slash as the same pattern', () => {
-    // `[-_/:]` is legal bare inside a character class and must be escaped
-    // outside one; that typesetting difference is not a divergence.
-    const bare = 'const RE = /a[-_/:]b/i;\n';
-    const escaped = 'const RE = /a[-_\\/:]b/i;\n';
-    expect(compareRegexDeclaration(bare, escaped, 'RE').match).toBe(true);
-  });
-
-  it('does not report a match when a side is missing or is not a regex', () => {
-    expect(compareRegexDeclaration(server, 'const other = 1;', 'TOOL_USE_RE').match).toBe(false);
-    expect(compareRegexDeclaration(server, 'const TOOL_USE_RE = 42;', 'TOOL_USE_RE').match).toBe(false);
-  });
-});
-
-describe('regexAlternationSource fails closed', () => {
-  // A scan-for-literals pass would silently DROP the non-literal element and
-  // report the surviving alternatives as the whole pattern — a mirror that has
-  // genuinely diverged then compares equal. Refuse the array instead.
-  it('refuses an array holding anything but string literals', () => {
-    expect(regexAlternationSource("const RE = new RegExp(['a', 'b'].join('|'), 'i');")).toBe('a|b');
-    expect(regexAlternationSource("const RE = new RegExp(['a', SHARED].join('|'), 'i');")).toBeNull();
-    expect(regexAlternationSource("const RE = new RegExp([...SHARED, 'b'].join('|'), 'i');")).toBeNull();
-  });
-
-  // `'\\x2e'` is the single character `.`, which matches ANYTHING — copying the
-  // escape through as the text `x2e` would compare equal to a client `/x2e/i`
-  // that means something entirely different.
-  it('refuses an escape it cannot decode by inspection', () => {
-    expect(regexAlternationSource("const RE = new RegExp(['a\\\\.b'].join('|'), 'i');")).toBe('a\\.b');
-    expect(regexAlternationSource("const RE = new RegExp(['\\x2e'].join('|'), 'i');")).toBeNull();
-    expect(regexAlternationSource("const RE = new RegExp(['\\n'].join('|'), 'i');")).toBeNull();
-  });
-
-  // A search-anywhere read would find the array and report the mirror intact
-  // while the declaration assigns something else entirely.
-  it('refuses a declaration that merely CONTAINS a recognizable form', () => {
-    const array = "new RegExp(['a', 'b'].join('|'), 'i')";
-    expect(regexAlternationSource(`const RE = ${array};`)).toBe('a|b');
-    expect(regexAlternationSource(`const RE = FLAG ? ${array} : OTHER_RE;`)).toBeNull();
-    expect(regexAlternationSource(`const RE = widen(${array});`)).toBeNull();
-  });
-
-  it('returns null for a declaration that is neither supported form', () => {
-    expect(regexAlternationSource(null)).toBeNull();
-    expect(regexAlternationSource('const RE = buildPattern();')).toBeNull();
   });
 });
