+++ server/lib/providerGateways.js
@@ -24,14 +24,13 @@
  * it imports nothing, so it stays importable from `providerModels.js`, which the
  * standalone autofixer process pulls in.
  *
- * **Three copies of this table exist, by architecture, and must stay in lockstep:**
+ * **Two copies of this table exist, by architecture, and must stay in lockstep:**
  *   1. this file — the PortOS server;
  *   2. `aiToolkit/internal/gateways.js` — the vendored toolkit, which may not
  *      import out of its own directory (see `aiToolkit/AGENTS.md`);
- *      `providerGateways.parity.test.js` fails when the two drift;
- *   3. `client/src/utils/providerGateways.js` — the browser, which cannot import server
- *      code at all; `providerGateways.parity.test.js` reads it as TEXT (never
- *      imports it) and pins the fields it carries.
+ *      `providerGateways.parity.test.js` fails when the two drift.
+ * The browser is not a third: this leaf is pure, so
+ * `client/src/utils/providerGateways.js` re-exports it.
  */
 
 /**
@@ -97,3 +96,6 @@ export function gatewayForProvider(provider) {
 
 /** `gatewayForProvider`, reduced to the id (which is also the OpenCode namespace). */
 export const gatewayIdForProvider = (provider) => gatewayForProvider(provider)?.id ?? null;
+
+/** True when a provider is an OpenCode wrapper front-ending any hosted gateway. */
+export const isGatewayBackedProvider = (provider) => gatewayForProvider(provider) !== null;
