--- server/services/stageRunner.mirror.test.js
@@ -1,71 +1,0 @@
-/**
- * Mirror parity test for the `KNOWN_MODEL_CONTEXT_WINDOWS` table, which lives
- * in both server/services/stageRunner.js and client/src/utils/providerContextWindows.js (the
- * client can't import server modules, so the table is duplicated behind a
- * "Keep in sync" comment on each side).
- *
- * Until now the mirror was pinned by that comment alone: each side had its own
- * hardcoded expectations (`stageRunner.test.js` / `providers.test.js`), so a
- * model bump that added a row to only one copy left BOTH suites green. The two
- * ends would then disagree about whether a given manuscript fits in one call —
- * the server chunking against the real window while the client's "context used"
- * meter falls through to the conservative DEFAULT_LARGE_CONTEXT_WINDOW, or vice
- * versa. Silent, and ~8x off.
- *
- * Only the table is mirrored: `knownModelContextWindow` is legitimately a
- * `function` server-side and an arrow const client-side, so it is not compared.
- * Comparison strips comments, so per-side commentary may diverge — logic can't.
- */
-
-import { describe, it, expect } from 'vitest';
-import { readFileSync } from 'fs';
-import { resolve, dirname } from 'path';
-import { fileURLToPath } from 'url';
-import { compareDeclaration } from '../lib/mirrorParity.js';
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
-const SERVER_PATH = resolve(__dirname, 'stageRunner.js');
-const CLIENT_PATH = resolve(__dirname, '../../client/src/utils/providerContextWindows.js');
-
-// The table's rows are `[/regex/i, SOME_CONTEXT_WINDOW]` — the *values* are
-// symbols defined separately in each file, so pinning the table text alone
-// leaves the exact hole this test exists to close: edit `CODEX_CONTEXT_WINDOW`
-// on one side only and the table still matches byte-for-byte while the two
-// ends silently disagree about the window. Pin the constants too.
-const MIRRORED_NAMES = [
-  'DEFAULT_LARGE_CONTEXT_WINDOW',
-  'CODEX_CONTEXT_WINDOW',
-  'GEMINI_CONTEXT_WINDOW',
-  'GROK_CONTEXT_WINDOW',
-  'KIMI_CONTEXT_WINDOW',
-  'KNOWN_MODEL_CONTEXT_WINDOWS',
-  // The catalog rung sits ABOVE the table in both ladders, so a divergence here
-  // is the same silent ~8x disagreement the table's parity closes. Both copies
-  // are declared `function` precisely so this comparison can be made.
-  'catalogModelContextWindow',
-];
-
-describe('stageRunner↔client providers context-window mirror parity', () => {
-  const serverSrc = readFileSync(SERVER_PATH, 'utf8');
-  const clientSrc = readFileSync(CLIENT_PATH, 'utf8');
-
-  it('both files are non-empty', () => {
-    expect(serverSrc.length).toBeGreaterThan(100);
-    expect(clientSrc.length).toBeGreaterThan(100);
-  });
-
-  for (const name of MIRRORED_NAMES) {
-    it(`${name} is present and identical on both sides (code only)`, () => {
-      const { serverDecl, clientDecl, serverNorm, clientNorm } =
-        compareDeclaration(serverSrc, clientSrc, name);
-
-      expect(serverDecl, `server/services/stageRunner.js is missing: ${name}`).not.toBeNull();
-      expect(clientDecl, `client/src/utils/providerContextWindows.js is missing: ${name}`).not.toBeNull();
-      expect(
-        clientNorm,
-        `"${name}" diverged — the server copy is authoritative; port the change verbatim`,
-      ).toBe(serverNorm);
-    });
-  }
-});
